Don't expose actual internal structs for benches, wrap them in special types
