+++ synedrion/benches/bench.rs
@@ -2,7 +2,9 @@ use criterion::{criterion_group, criterion_main, Criterion};
 use rand_core::OsRng;
 
 use synedrion::{
-    bench_internals::{benches, KeyShare, PresigningData},
+    bench_internals::{
+        key_init, key_refresh, presigning, signing, PresigningInputs, SigningInputs,
+    },
     TestParams,
 };
 
@@ -10,24 +12,25 @@ fn bench_happy_paths(c: &mut Criterion) {
     let mut group = c.benchmark_group("happy path");
 
     type Params = TestParams;
-    let key_shares = KeyShare::new_centralized(&mut OsRng, 2, None);
 
     group.bench_function("KeyGen, 2 parties", |b| {
-        b.iter(|| benches::key_init::<Params>(&mut OsRng, 2))
+        b.iter(|| key_init::<Params>(&mut OsRng, 2))
     });
 
-    let presigning_datas = PresigningData::new_centralized(&mut OsRng, &key_shares);
+    let presigning_inputs = PresigningInputs::new(&mut OsRng, 2);
+    let signing_inputs = SigningInputs::new(&mut OsRng, &presigning_inputs);
+
     group.bench_function("Signing, 2 parties", |b| {
-        b.iter(|| benches::signing::<Params>(&mut OsRng, &key_shares, &presigning_datas))
+        b.iter(|| signing::<Params>(&mut OsRng, &presigning_inputs, &signing_inputs))
     });
 
     group.sample_size(10);
     group.bench_function("Presigning, 2 parties", |b| {
-        b.iter(|| benches::presigning::<Params>(&mut OsRng, &key_shares))
+        b.iter(|| presigning::<Params>(&mut OsRng, &presigning_inputs))
     });
 
     group.bench_function("KeyRefresh, 2 parties", |b| {
-        b.iter(|| benches::key_refresh::<Params>(&mut OsRng, 2))
+        b.iter(|| key_refresh::<Params>(&mut OsRng, 2))
     });
 
     group.finish()
